refactor: 핵심 비즈니스 로직을 가지는 객체를 domain 패키지에, UI 관련 객체를 view 패키지로 이동
