--- src/main/java/racingcar/domain/input/carname/printer/CarNamesInputFromUserPrinter.java
@@ -1,7 +1,0 @@
-package racingcar.domain.input.carname.printer;
-
-public class CarNamesInputFromUserPrinter {
-    public void printCarNamesInputMessage() {
-        System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
-    }
-}
