--- src/main/java/racingcar/domain/input/carname/validator/exception/LessThanTwoCarNamesException.java
@@ -1,7 +1,0 @@
-package racingcar.domain.input.carname.validator.exception;
-
-public class LessThanTwoCarNamesException extends RuntimeException {
-    public LessThanTwoCarNamesException(String message) {
-        super(message);
-    }
-}
