--- src/main/java/racingcar/domain/input/racingtrytime/RawRacingTryTimeInputFromUser.java
@@ -1,19 +1,0 @@
-package racingcar.domain.input.racingtrytime;
-
-import java.util.Scanner;
-import racingcar.domain.input.racingtrytime.printer.RacingTryTimeInputFromUserPrinter;
-
-public class RawRacingTryTimeInputFromUser {
-    private final RacingTryTimeInputFromUserPrinter racingTryTimeInputFromUserPrinter;
-    private final Scanner scanner;
-
-    public RawRacingTryTimeInputFromUser(Scanner scanner) {
-        this.racingTryTimeInputFromUserPrinter = new RacingTryTimeInputFromUserPrinter();
-        this.scanner = scanner;
-    }
-
-    public String getRacingTryTimeInput() {
-        racingTryTimeInputFromUserPrinter.printRacingTryTimeInputMessage();
-        return scanner.nextLine();
-    }
-}
