--- src/main/java/racingcar/domain/input/racingtrytime/validator/exception/NotNaturalNumberRacingTimeException.java
@@ -1,7 +1,0 @@
-package racingcar.domain.input.racingtrytime.validator.exception;
-
-public class NotNaturalNumberRacingTimeException extends RuntimeException {
-    public NotNaturalNumberRacingTimeException(String message) {
-        super(message);
-    }
-}
