+++ src/main/java/racingcar/domain/racing/CarRacing.java
@@ -3,7 +3,7 @@ package racingcar.domain.racing;
 
 import racingcar.domain.CarRepository;
 import racingcar.domain.numbergenerator.NumberGenerator;
-import racingcar.domain.output.RacingPrinter;
+import racingcar.view.output.RacingPrinter;
 
 public class CarRacing {
     private final NumberApplicatorToCar applicatorRandomNumberToCar;
