+++ src/main/java/racingcar/view/input/ValidInputFromUserGetter.java
@@ -1,0 +1,24 @@
+package racingcar.view.input;
+
+import java.util.List;
+import java.util.Scanner;
+import racingcar.view.input.carname.ValidCarNamesInputFromUserGetter;
+import racingcar.view.input.racingtrytime.ValidRacingTryTimeFromUserGetter;
+
+public class ValidInputFromUserGetter {
+    private final ValidCarNamesInputFromUserGetter validCarNamesInputFromUserGetter;
+    private final ValidRacingTryTimeFromUserGetter validRacingTryTimeFromUserGetter;
+
+    public ValidInputFromUserGetter(Scanner scanner) {
+        this.validCarNamesInputFromUserGetter = new ValidCarNamesInputFromUserGetter(scanner);
+        this.validRacingTryTimeFromUserGetter = new ValidRacingTryTimeFromUserGetter(scanner);
+    }
+
+    public List<String> getCarNames() {
+        return validCarNamesInputFromUserGetter.getCarNames();
+    }
+
+    public int getRacingTryTime() {
+        return validRacingTryTimeFromUserGetter.getRacingTryTime();
+    }
+}
