+++ src/main/java/racingcar/view/input/carname/validator/exception/DuplicateCarNamesException.java
@@ -1,0 +1,7 @@
+package racingcar.view.input.carname.validator.exception;
+
+public class DuplicateCarNamesException extends RuntimeException {
+    public DuplicateCarNamesException(String message) {
+        super(message);
+    }
+}
