+++ src/main/java/racingcar/view/input/carname/validator/exception/InvalidCarNameFormatException.java
@@ -1,0 +1,7 @@
+package racingcar.view.input.carname.validator.exception;
+
+public class InvalidCarNameFormatException extends RuntimeException {
+    public InvalidCarNameFormatException(String message) {
+        super(message);
+    }
+}
