+++ src/main/java/racingcar/view/input/carname/validator/exception/LessThanTwoCarNamesException.java
@@ -1,0 +1,7 @@
+package racingcar.view.input.carname.validator.exception;
+
+public class LessThanTwoCarNamesException extends RuntimeException {
+    public LessThanTwoCarNamesException(String message) {
+        super(message);
+    }
+}
