+++ src/main/java/racingcar/view/input/racingtrytime/validator/exception/NotNaturalNumberRacingTimeException.java
@@ -1,0 +1,7 @@
+package racingcar.view.input.racingtrytime.validator.exception;
+
+public class NotNaturalNumberRacingTimeException extends RuntimeException {
+    public NotNaturalNumberRacingTimeException(String message) {
+        super(message);
+    }
+}
